@@ -12,7 +12,7 @@ import { LoaderService } from './loader.service';
   styleUrls: ['./car-presenter.component.css']
 })
 export class CarPresenterComponent implements OnInit {
-  @ViewChild('stage') stageEl;
+  @ViewChild('stage', { static: true }) stageEl;
   scene: THREE.Scene;
   camera: THREE.PerspectiveCamera;
   renderer: THREE.WebGLRenderer;
